Do not attempt to remove a file that is not clean

+++ plugins/git/include/Git_Exec.class.php
@@ -69,8 +69,17 @@ class Git_Exec {
      * @throw Git_Command_Exception
      */
     public function rm($file) {
-        $cmd = 'git rm '.escapeshellarg($file);
-        return $this->gitCmd($cmd);
+        if ($this->canRemove($file)) {
+            $cmd = 'git rm '.escapeshellarg($file);
+            return $this->gitCmd($cmd);
+        }
+        return true;
+    }
+
+    private function canRemove($file) {
+        $output = array();
+        $this->execInPath('git status --porcelain '.escapeshellarg($file), $output);
+        return count($output) == 0;
     }
 
     /**
